Refactor the data model, Doctrine, and validation structures
