+++ Resources/config/doctrine/RefreshToken.orm.xml
@@ -5,9 +5,13 @@
                   xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping
                     http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
 
-    <mapped-superclass name="Gesdinet\JWTRefreshTokenBundle\Entity\RefreshToken">
+    <entity name="Gesdinet\JWTRefreshTokenBundle\Entity\RefreshToken" table="refresh_tokens">
         <id name="id" type="integer">
             <generator strategy="AUTO"/>
         </id>
-    </mapped-superclass>
+
+        <field name="refreshToken" type="string" column="refresh_token" length="128" unique="true"/>
+        <field name="username" type="string" length="255" column="username"/>
+        <field name="valid" type="datetime"/>
+    </entity>
 </doctrine-mapping>
