--- Resources/config/mongodb/doctrine-superclass/RefreshToken.mongodb.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<doctrine-mongo-mapping xmlns="http://doctrine-project.org/schemas/odm/doctrine-mongo-mapping"
-                        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-                        xsi:schemaLocation="http://doctrine-project.org/schemas/odm/doctrine-mongo-mapping
-                    http://doctrine-project.org/schemas/odm/doctrine-mongo-mapping.xsd">
-
-    <mapped-superclass name="Gesdinet\JWTRefreshTokenBundle\Document\RefreshToken">
-        <field field-name="id" type="id" />
-    </mapped-superclass>
-</doctrine-mongo-mapping>
