--- Resources/config/orm/doctrine-orm/AbstractRefreshToken.orm.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping"
-                  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-                  xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping
-                    http://doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
-
-    <mapped-superclass name="Gesdinet\JWTRefreshTokenBundle\Entity\AbstractRefreshToken">
-        <field name="refreshToken" type="string" column="refresh_token" length="128" unique="true"/>
-        <field name="username" type="string" length="255" column="username"/>
-        <field name="valid" type="datetime"/>
-    </mapped-superclass>
-</doctrine-mapping>
