+++ Resources/config/validation/AbstractRefreshToken.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<constraint-mapping xmlns="http://symfony.com/schema/dic/constraint-mapping"
+                    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+                    xsi:schemaLocation="http://symfony.com/schema/dic/constraint-mapping http://symfony.com/schema/dic/services/constraint-mapping-1.0.xsd">
+    <class name="Gesdinet\JWTRefreshTokenBundle\Model\AbstractRefreshToken">
+        <property name="refreshToken">
+            <constraint name="NotBlank" />
+        </property>
+        <property name="username">
+            <constraint name="NotBlank" />
+        </property>
+        <property name="valid">
+            <constraint name="NotBlank" />
+        </property>
+    </class>
+</constraint-mapping>
